+++ test/unit/argument_parser/format_parse_validators_test.cpp
@@ -136,7 +136,7 @@ TEST(validator_test, input_file)
         seqan3::argument_parser parser{"test_parser", 3, argv, seqan3::update_notifications::off};
         test_accessor::set_terminal_width(parser, 80);
         parser.add_option(file_in_path, 'i', "int-option", "desc",
-                          seqan3::option_spec::DEFAULT, seqan3::input_file_validator{formats});
+                          seqan3::option_spec::standard, seqan3::input_file_validator{formats});
 
         EXPECT_NO_THROW(parser.parse());
         EXPECT_EQ(file_in_path.string(), path);
@@ -254,7 +254,7 @@ TEST(validator_test, output_file)
         seqan3::argument_parser parser{"test_parser", 3, argv, seqan3::update_notifications::off};
         test_accessor::set_terminal_width(parser, 80);
         parser.add_option(file_out_path, 'o', "out-option", "desc",
-                          seqan3::option_spec::DEFAULT,
+                          seqan3::option_spec::standard,
                           seqan3::output_file_validator{seqan3::output_file_open_options::create_new, formats});
 
         EXPECT_NO_THROW(parser.parse());
@@ -389,7 +389,7 @@ TEST(validator_test, input_directory)
             seqan3::argument_parser parser{"test_parser", 3, argv, seqan3::update_notifications::off};
             test_accessor::set_terminal_width(parser, 80);
             parser.add_option(dir_in_path, 'i', "input-option", "desc",
-                              seqan3::option_spec::DEFAULT, seqan3::input_directory_validator{});
+                              seqan3::option_spec::standard, seqan3::input_directory_validator{});
 
             EXPECT_NO_THROW(parser.parse());
             EXPECT_EQ(path, dir_in_path.string());
@@ -441,7 +441,7 @@ TEST(validator_test, output_directory)
         seqan3::argument_parser parser{"test_parser", 3, argv, seqan3::update_notifications::off};
         test_accessor::set_terminal_width(parser, 80);
         parser.add_option(dir_out_path, 'o', "output-option", "desc",
-                          seqan3::option_spec::DEFAULT,
+                          seqan3::option_spec::standard,
                           seqan3::output_directory_validator{});
 
         EXPECT_NO_THROW(parser.parse());
@@ -642,7 +642,7 @@ TEST(validator_test, arithmetic_range_validator_success)
     seqan3::argument_parser parser{"test_parser", 3, argv, seqan3::update_notifications::off};
     test_accessor::set_terminal_width(parser, 80);
     parser.add_option(option_value, 'i', "int-option", "desc",
-                      seqan3::option_spec::DEFAULT, seqan3::arithmetic_range_validator{1, 20});
+                      seqan3::option_spec::standard, seqan3::arithmetic_range_validator{1, 20});
 
     testing::internal::CaptureStderr();
     EXPECT_NO_THROW(parser.parse());
@@ -654,7 +654,7 @@ TEST(validator_test, arithmetic_range_validator_success)
     seqan3::argument_parser parser2{"test_parser", 3, argv2, seqan3::update_notifications::off};
     test_accessor::set_terminal_width(parser2, 80);
     parser2.add_option(option_value, 'i', "int-option", "desc",
-                       seqan3::option_spec::DEFAULT, seqan3::arithmetic_range_validator{-20, 20});
+                       seqan3::option_spec::standard, seqan3::arithmetic_range_validator{-20, 20});
 
     testing::internal::CaptureStderr();
     EXPECT_NO_THROW(parser2.parse());
@@ -688,7 +688,7 @@ TEST(validator_test, arithmetic_range_validator_success)
     seqan3::argument_parser parser5{"test_parser", 5, argv5, seqan3::update_notifications::off};
     test_accessor::set_terminal_width(parser5, 80);
     parser5.add_option(option_vector, 'i', "int-option", "desc",
-                       seqan3::option_spec::DEFAULT, seqan3::arithmetic_range_validator{-50,50});
+                       seqan3::option_spec::standard, seqan3::arithmetic_range_validator{-50,50});
 
     testing::internal::CaptureStderr();
     EXPECT_NO_THROW(parser5.parse());
@@ -737,7 +737,7 @@ TEST(validator_test, arithmetic_range_validator_success)
     seqan3::argument_parser parser8{"test_parser", 3, argv8, seqan3::update_notifications::off};
     test_accessor::set_terminal_width(parser8, 80);
     parser8.add_option(double_option_value, 'i', "double-option", "desc",
-                       seqan3::option_spec::DEFAULT, seqan3::arithmetic_range_validator{1, 20});
+                       seqan3::option_spec::standard, seqan3::arithmetic_range_validator{1, 20});
 
     testing::internal::CaptureStderr();
     EXPECT_NO_THROW(parser8.parse());
@@ -755,7 +755,7 @@ TEST(validator_test, arithmetic_range_validator_error)
     seqan3::argument_parser parser{"test_parser", 3, argv, seqan3::update_notifications::off};
     test_accessor::set_terminal_width(parser, 80);
     parser.add_option(option_value, 'i', "int-option", "desc",
-                      seqan3::option_spec::DEFAULT, seqan3::arithmetic_range_validator{1, 20});
+                      seqan3::option_spec::standard, seqan3::arithmetic_range_validator{1, 20});
 
     EXPECT_THROW(parser.parse(), seqan3::validation_error);
 
@@ -764,7 +764,7 @@ TEST(validator_test, arithmetic_range_validator_error)
     seqan3::argument_parser parser2{"test_parser", 3, argv2, seqan3::update_notifications::off};
     test_accessor::set_terminal_width(parser2, 80);
     parser2.add_option(option_value, 'i', "int-option", "desc",
-                       seqan3::option_spec::DEFAULT, seqan3::arithmetic_range_validator{-20, 20});
+                       seqan3::option_spec::standard, seqan3::arithmetic_range_validator{-20, 20});
 
     EXPECT_THROW(parser2.parse(), seqan3::validation_error);
 
@@ -789,7 +789,7 @@ TEST(validator_test, arithmetic_range_validator_error)
     seqan3::argument_parser parser5{"test_parser", 3, argv5, seqan3::update_notifications::off};
     test_accessor::set_terminal_width(parser5, 80);
     parser5.add_option(option_vector, 'i', "int-option", "desc",
-                       seqan3::option_spec::DEFAULT, seqan3::arithmetic_range_validator{-50, 50});
+                       seqan3::option_spec::standard, seqan3::arithmetic_range_validator{-50, 50});
 
     EXPECT_THROW(parser5.parse(), seqan3::validation_error);
 
@@ -808,7 +808,7 @@ TEST(validator_test, arithmetic_range_validator_error)
     seqan3::argument_parser parser7{"test_parser", 3, argv7, seqan3::update_notifications::off};
     test_accessor::set_terminal_width(parser7, 80);
     parser7.add_option(double_option_value, 'i', "double-option", "desc",
-                       seqan3::option_spec::DEFAULT, seqan3::arithmetic_range_validator{1, 20});
+                       seqan3::option_spec::standard, seqan3::arithmetic_range_validator{1, 20});
 
     EXPECT_THROW(parser7.parse(), seqan3::validation_error);
 }
@@ -868,7 +868,7 @@ TEST(validator_test, value_list_validator_success)
     seqan3::argument_parser parser{"test_parser", 3, argv, seqan3::update_notifications::off};
     test_accessor::set_terminal_width(parser, 80);
     parser.add_option(option_value, 's', "string-option", "desc",
-                      seqan3::option_spec::DEFAULT,
+                      seqan3::option_spec::standard,
                       seqan3::value_list_validator{valid_str_values | std::views::take(2)});
 
     testing::internal::CaptureStderr();
@@ -881,7 +881,7 @@ TEST(validator_test, value_list_validator_success)
     seqan3::argument_parser parser2{"test_parser", 3, argv2, seqan3::update_notifications::off};
     test_accessor::set_terminal_width(parser2, 80);
     parser2.add_option(option_value_int, 'i', "int-option", "desc",
-                       seqan3::option_spec::DEFAULT, seqan3::value_list_validator<int>{0, -21, 10});
+                       seqan3::option_spec::standard, seqan3::value_list_validator<int>{0, -21, 10});
 
     testing::internal::CaptureStderr();
     EXPECT_NO_THROW(parser2.parse());
@@ -916,7 +916,7 @@ TEST(validator_test, value_list_validator_success)
     seqan3::argument_parser parser5{"test_parser", 5, argv5, seqan3::update_notifications::off};
     test_accessor::set_terminal_width(parser5, 80);
     parser5.add_option(option_vector_int, 'i', "int-option", "desc",
-                       seqan3::option_spec::DEFAULT, seqan3::value_list_validator<int>{-10, 48, 50});
+                       seqan3::option_spec::standard, seqan3::value_list_validator<int>{-10, 48, 50});
 
     testing::internal::CaptureStderr();
     EXPECT_NO_THROW(parser5.parse());
@@ -930,7 +930,7 @@ TEST(validator_test, value_list_validator_success)
     seqan3::argument_parser parser7{"test_parser", 2, argv7, seqan3::update_notifications::off};
     test_accessor::set_terminal_width(parser7, 80);
     parser7.add_option(option_vector_int, 'i', "int-option", "desc",
-                       seqan3::option_spec::DEFAULT, seqan3::value_list_validator<int>{-10, 48, 50});
+                       seqan3::option_spec::standard, seqan3::value_list_validator<int>{-10, 48, 50});
 
     option_vector_int.clear();
     testing::internal::CaptureStdout();
@@ -958,7 +958,7 @@ TEST(validator_test, value_list_validator_error)
     seqan3::argument_parser parser{"test_parser", 3, argv, seqan3::update_notifications::off};
     test_accessor::set_terminal_width(parser, 80);
     parser.add_option(option_value, 's', "string-option", "desc",
-                      seqan3::option_spec::DEFAULT, seqan3::value_list_validator{"ha", "ba", "ma"});
+                      seqan3::option_spec::standard, seqan3::value_list_validator{"ha", "ba", "ma"});
 
     EXPECT_THROW(parser.parse(), seqan3::validation_error);
 
@@ -984,7 +984,7 @@ TEST(validator_test, value_list_validator_error)
     seqan3::argument_parser parser5{"test_parser", 5, argv5, seqan3::update_notifications::off};
     test_accessor::set_terminal_width(parser5, 80);
     parser5.add_option(option_vector_int, 'i', "int-option", "desc",
-                       seqan3::option_spec::DEFAULT, seqan3::value_list_validator<int>{-10, 48, 50});
+                       seqan3::option_spec::standard, seqan3::value_list_validator<int>{-10, 48, 50});
 
     EXPECT_THROW(parser5.parse(), seqan3::validation_error);
 }
@@ -1001,7 +1001,7 @@ TEST(validator_test, regex_validator_success)
         seqan3::argument_parser parser{"test_parser", 3, argv, seqan3::update_notifications::off};
         test_accessor::set_terminal_width(parser, 80);
         parser.add_option(option_value, 's', "string-option", "desc",
-                          seqan3::option_spec::DEFAULT, email_validator);
+                          seqan3::option_spec::standard, email_validator);
 
         testing::internal::CaptureStderr();
         EXPECT_NO_THROW(parser.parse());
@@ -1043,7 +1043,7 @@ TEST(validator_test, regex_validator_success)
         seqan3::argument_parser parser{"test_parser", 5, argv, seqan3::update_notifications::off};
         test_accessor::set_terminal_width(parser, 80);
         parser.add_option(option_vector, 's', "string-option", "desc",
-                           seqan3::option_spec::DEFAULT, email_vector_validator);
+                           seqan3::option_spec::standard, email_vector_validator);
 
         testing::internal::CaptureStderr();
         EXPECT_NO_THROW(parser.parse());
@@ -1058,7 +1058,7 @@ TEST(validator_test, regex_validator_success)
         seqan3::argument_parser parser{"test_parser", 3, argv, seqan3::update_notifications::off};
         test_accessor::set_terminal_width(parser, 80);
         parser.add_option(path_option, 's', "string-option", "desc",
-                          seqan3::option_spec::DEFAULT, email_vector_validator);
+                          seqan3::option_spec::standard, email_vector_validator);
 
         testing::internal::CaptureStderr();
         EXPECT_NO_THROW(parser.parse());
@@ -1072,7 +1072,7 @@ TEST(validator_test, regex_validator_success)
         seqan3::argument_parser parser{"test_parser", 2, argv, seqan3::update_notifications::off};
         test_accessor::set_terminal_width(parser, 80);
         parser.add_option(option_vector, 's', "string-option", "desc",
-                           seqan3::option_spec::DEFAULT, email_vector_validator);
+                           seqan3::option_spec::standard, email_vector_validator);
 
         option_vector.clear();
         testing::internal::CaptureStdout();
@@ -1101,7 +1101,7 @@ TEST(validator_test, regex_validator_error)
     seqan3::argument_parser parser{"test_parser", 3, argv, seqan3::update_notifications::off};
     test_accessor::set_terminal_width(parser, 80);
     parser.add_option(option_value, '\0', "string-option", "desc",
-                      seqan3::option_spec::DEFAULT, seqan3::regex_validator{"tt"});
+                      seqan3::option_spec::standard, seqan3::regex_validator{"tt"});
 
     EXPECT_THROW(parser.parse(), seqan3::validation_error);
 
@@ -1129,7 +1129,7 @@ TEST(validator_test, regex_validator_error)
     seqan3::argument_parser parser4{"test_parser", 5, argv4, seqan3::update_notifications::off};
     test_accessor::set_terminal_width(parser4, 80);
     parser4.add_option(option_vector, 's', "", "desc",
-                       seqan3::option_spec::DEFAULT, seqan3::regex_validator{"tt"});
+                       seqan3::option_spec::standard, seqan3::regex_validator{"tt"});
 
     EXPECT_THROW(parser4.parse(), seqan3::validation_error);
 }
@@ -1152,7 +1152,7 @@ TEST(validator_test, chaining_validators)
         seqan3::argument_parser parser{"test_parser", 3, argv, seqan3::update_notifications::off};
         test_accessor::set_terminal_width(parser, 80);
         parser.add_option(option_value, 's', "string-option", "desc",
-                          seqan3::option_spec::DEFAULT, absolute_path_validator | my_file_ext_validator);
+                          seqan3::option_spec::standard, absolute_path_validator | my_file_ext_validator);
 
         testing::internal::CaptureStderr();
         EXPECT_NO_THROW(parser.parse());
@@ -1166,7 +1166,7 @@ TEST(validator_test, chaining_validators)
         seqan3::argument_parser parser{"test_parser", 3, argv, seqan3::update_notifications::off};
         test_accessor::set_terminal_width(parser, 80);
         parser.add_option(option_value, 's', "string-option", "desc",
-                          seqan3::option_spec::DEFAULT, absolute_path_validator | my_file_ext_validator);
+                          seqan3::option_spec::standard, absolute_path_validator | my_file_ext_validator);
 
         EXPECT_THROW(parser.parse(), seqan3::validation_error);
     }
@@ -1177,7 +1177,7 @@ TEST(validator_test, chaining_validators)
         seqan3::argument_parser parser{"test_parser", 3, argv, seqan3::update_notifications::off};
         test_accessor::set_terminal_width(parser, 80);
         parser.add_option(option_value, 's', "string-option", "desc",
-                          seqan3::option_spec::DEFAULT, absolute_path_validator | my_file_ext_validator);
+                          seqan3::option_spec::standard, absolute_path_validator | my_file_ext_validator);
 
         EXPECT_THROW(parser.parse(), seqan3::validation_error);
     }
@@ -1189,7 +1189,7 @@ TEST(validator_test, chaining_validators)
         seqan3::argument_parser parser{"test_parser", 3, argv, seqan3::update_notifications::off};
         test_accessor::set_terminal_width(parser, 80);
         parser.add_option(option_value, 's', "string-option", "desc",
-                          seqan3::option_spec::DEFAULT,
+                          seqan3::option_spec::standard,
                           seqan3::regex_validator{"(/[^/]+)+/.*\\.[^/\\.]+$"} |
                           seqan3::output_file_validator{seqan3::output_file_open_options::create_new, {"sa", "so"}});
 
@@ -1206,7 +1206,7 @@ TEST(validator_test, chaining_validators)
         seqan3::argument_parser parser{"test_parser", 3, argv, seqan3::update_notifications::off};
         test_accessor::set_terminal_width(parser, 80);
         parser.add_option(option_value, 's', "string-option", "desc",
-                          seqan3::option_spec::DEFAULT,
+                          seqan3::option_spec::standard,
                           seqan3::regex_validator{"(/[^/]+)+/.*\\.[^/\\.]+$"} |
                           seqan3::output_file_validator{seqan3::output_file_open_options::create_new, {"sa", "so"}} |
                           seqan3::regex_validator{".*"});
@@ -1224,7 +1224,7 @@ TEST(validator_test, chaining_validators)
         seqan3::argument_parser parser{"test_parser", 2, argv, seqan3::update_notifications::off};
         test_accessor::set_terminal_width(parser, 80);
         parser.add_option(option_value, 's', "string-option", "desc",
-                          seqan3::option_spec::DEFAULT,
+                          seqan3::option_spec::standard,
                           seqan3::regex_validator{"(/[^/]+)+/.*\\.[^/\\.]+$"} |
                           seqan3::output_file_validator{seqan3::output_file_open_options::create_new, {"sa", "so"}} |
                           seqan3::regex_validator{".*"});
@@ -1253,7 +1253,7 @@ TEST(validator_test, chaining_validators)
         seqan3::argument_parser parser{"test_parser", 2, argv, seqan3::update_notifications::off};
         test_accessor::set_terminal_width(parser, 80);
         parser.add_option(option_value, 's', "string-option", "desc",
-                          seqan3::option_spec::DEFAULT,
+                          seqan3::option_spec::standard,
                           seqan3::regex_validator{"(/[^/]+)+/.*\\.[^/\\.]+$"} |
                           seqan3::output_file_validator{seqan3::output_file_open_options::open_or_create, {"sa", "so"}} |
                           seqan3::regex_validator{".*"});
@@ -1282,7 +1282,7 @@ TEST(validator_test, chaining_validators)
         seqan3::argument_parser parser{"test_parser", 3, argv, seqan3::update_notifications::off};
         test_accessor::set_terminal_width(parser, 80);
         parser.add_option(option_list_value, 's', "string-option", "desc",
-                          seqan3::option_spec::DEFAULT,
+                          seqan3::option_spec::standard,
                           seqan3::regex_validator{"(/[^/]+)+/.*\\.[^/\\.]+$"} |
                           seqan3::output_file_validator{seqan3::output_file_open_options::create_new, {"sa", "so"}});